vertex-ai-automl-imageclassification-webapp/requirements.txt
gunicorn
flask
requests
google-auth
google-cloud-aiplatform
protobuf